requirements.txt
requests==2.31.0
sqlalchemy==2.0.25
pymysql==1.1.0
python-dotenv==1.0.1
pyyaml==6.0.1